requirements.txt
gradio>=4.44.0
PyMuPDF>=1.23.0
deepseek-ocr>=0.1.1
sentence-transformers>=2.2.2
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
scikit-learn>=1.3.0
layoutparser>=0.3.4
camelot-py>=0.11.0
pdfplumber>=0.11.0
reportlab>=4.0.0
pydantic>=1.10.0
pytest>=8.0.0
transformers==4.46.3
easydict>=1.10
torch>=2.0.0
huggingface_hub>=0.20.0
addict>=2.4.0
matplotlib>=3.10.0
torchvision>=0.24.0
einops>=0.7.0
dtaidistance>=2.3.0
beautifulsoup4>=4.12.0
rapidfuzz>=3.0.0
paddleocr>=2.7.0
pytesseract>=0.3.10
pdf2image>=1.17.0
jiwer>=3.0.5
ultralytics>=8.3.0
segment-anything>=1.0
doclayout-yolo>=0.0.4
pydantic-settings>=2.0.3
imagehash>=4.3.0

# Utilities and Integration
psutil>=5.9.0
requests>=2.31.0
sseclient-py>=1.8.0
tqdm>=4.66.0
scipy>=1.11.0
pytest-cov>=4.1.0
